+++ modules/product/infrastructure/product_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Boolean
+from sqlalchemy import Column, Integer, String, Boolean, Float
 from config import Base
 
 class Product(Base):
@@ -7,5 +7,5 @@ class Product(Base):
     id = Column(Integer, primary_key=True, index=True)
     name = Column(String, index=True)
     description = Column(String)
-    price = Column(Integer)
+    price = Column(Float)
     is_active = Column(Boolean, default=True)
